completed component structure
