+++ app/restaurant/[slug]/components/MenuCard.tsx
@@ -1,0 +1,11 @@
+export default function MenuCard() {
+  return (
+    <div className=" border rounded p-3 w-[49%] mb-3">
+      <h3 className="font-bold text-lg">Surf and Turf</h3>
+      <p className="font-light mt-1 text-sm">
+        A well done steak with lobster and rice
+      </p>
+      <p className="mt-7">$80.00</p>
+    </div>
+  );
+}
